backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14,<3.16
djangorestframework-simplejwt>=5.3,<5.5
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
gunicorn>=22.0.0
